requirements.txt
fastapi==0.103.1
line-bot-sdk==3.5.0
python-dotenv==1.0.0
uvicorn==0.23.2
aiohttp==3.8.5
pydantic==2.3.0